Updating title string generation in engine.
The engine class now contains the logic to generate the title string.
That means the GUI now calls a function to get the title string instead
of generating it everytime it is needed.  This required updating the
width and hieght to f32 from u32 so the engine can be generated before
the nannou window is created.  The test-files/glider_test.txt had the
step and update rates removed to test different title configurations.